+++ product-service/src/constants/common.ts
@@ -23,22 +23,9 @@ export enum HttpMethod {
   PUT = 'PUT',
 }
 
-export type Res = {
-  statusCode: number;
-  headers: Record<string, any>;
-  body: string;
-};
-
 export enum ErrMsg {
   INVALID_HTTP_METHOD = 'Invalid HTTP method',
   NOT_INSTANCE_OF_ERROR = 'Failed to do something exceptional',
   MISSING_ID = 'Missing path parameter: id',
   DOES_NOT_EXIST = `Product with such id doesn't exist`,
 }
-
-export type Product = {
-  id: number;
-  title: string;
-  description: string;
-  price: number;
-};
