+++ product-service/src/db/index.ts
@@ -1,0 +1,8 @@
+export { default as db } from './mockDb.json';
+
+export type Product = {
+  id: number;
+  title: string;
+  description: string;
+  price: number;
+};
