+++ product-service/src/types/index.ts
@@ -1,0 +1,5 @@
+export type Res = {
+  statusCode: number;
+  headers: Record<string, any>;
+  body: string;
+};
