+++ product-service/src/utils/db.ts
@@ -1,5 +1,4 @@
-import { Product } from '../constants';
-import { default as db } from '../db/mockDb.json';
+import { Product, db } from '../db';
 
 export const getOne = ({ id }: { id: string }): Product | undefined => {
   return db.find((it) => it.id === parseInt(id));
